Move ROS2 db3 converter tests
